separate generic tests and address comments
