+++ captum/_utils/common.py
@@ -5,7 +5,8 @@ from inspect import signature
 from typing import Any, Callable, List, Tuple, Union, cast, overload
 
 import torch
-from torch import Tensor
+from torch import Tensor, device
+from torch.nn import Module
 
 from .._utils.typing import Literal, TargetType
 
@@ -13,6 +14,14 @@ from .._utils.typing import Literal, TargetType
 class ExpansionTypes(Enum):
     repeat = 1
     repeat_interleave = 2
+
+
+def safe_div(denom: Tensor, quotient: float, default_value: Tensor) -> Tensor:
+    r"""
+        A simple utility function to perform `denom / quotient`
+        if the statement is undefined => result will be `default_value`
+    """
+    return denom / quotient if quotient != 0.0 else default_value
 
 
 @typing.overload
@@ -216,3 +225,34 @@ def _verify_select_column(
         len(target) <= len(output.shape) - 1
     ), "Cannot choose target column with output shape %r." % (output.shape,)
     return output[(slice(None), *target)]
+
+
+def _extract_device(
+    module: Module,
+    hook_inputs: Union[None, Tensor, Tuple[Tensor, ...]],
+    hook_outputs: Union[None, Tensor, Tuple[Tensor, ...]],
+) -> device:
+    params = list(module.parameters())
+    if (
+        (hook_inputs is None or len(hook_inputs) == 0)
+        and (hook_outputs is None or len(hook_outputs) == 0)
+        and len(params) == 0
+    ):
+        raise RuntimeError(
+            """Unable to extract device information for the module
+            {}. Both inputs and outputs to the forward hook and
+            `module.parameters()` are empty.
+            The reason that the inputs to the forward hook are empty
+            could be due to the fact that the arguments to that
+            module {} are all named and are passed as named
+            variables to its forward function.
+            """.format(
+                module, module
+            )
+        )
+    if hook_inputs is not None and len(hook_inputs) > 0:
+        return hook_inputs[0].device
+    if hook_outputs is not None and len(hook_outputs) > 0:
+        return hook_outputs[0].device
+
+    return params[0].device
