+++ captum/attr/_core/layer/layer_integrated_gradients.py
@@ -10,7 +10,6 @@ from torch.nn.parallel.scatter_gather import scatter
 from captum.attr._core.integrated_gradients import IntegratedGradients
 from captum.attr._utils.attribution import GradientAttribution, LayerAttribution
 from captum.attr._utils.common import (
-    _extract_device,
     _format_attributions,
     _format_input_baseline,
     _tensorize_baseline,
@@ -18,7 +17,7 @@ from captum.attr._utils.common import (
 )
 from captum.attr._utils.gradient import _forward_layer_eval, _run_forward
 
-from ...._utils.common import _format_additional_forward_args
+from ...._utils.common import _extract_device, _format_additional_forward_args
 from ...._utils.typing import BaselineType, Literal, TargetType
