+++ captum/insights/features.py
@@ -7,8 +7,8 @@ from typing import Callable, List, Optional, Union
 
 import numpy as np
 
+from captum._utils.common import safe_div
 from captum.attr._utils import visualization as viz
-from captum.attr._utils.common import safe_div
 from captum.insights._utils.transforms import format_transforms
 
 FeatureOutput = namedtuple("FeatureOutput", "name base modified type contribution")
