+++ tests/attr/layer/test_grad_cam.py
@@ -9,8 +9,8 @@ from torch.nn import Module
 
 from captum.attr._core.layer.grad_cam import LayerGradCam
 
+from ...helpers.basic import BaseTest, assertTensorTuplesAlmostEqual
 from ..helpers.basic_models import BasicModel_ConvNet_One_Conv, BasicModel_MultiLayer
-from ..helpers.utils import BaseTest, assertTensorTuplesAlmostEqual
 
 
 class Test(BaseTest):
