+++ tests/attr/layer/test_layer_conductance.py
@@ -10,17 +10,17 @@ from torch.nn import Module
 from captum._utils.typing import BaselineType
 from captum.attr._core.layer.layer_conductance import LayerConductance
 
+from ...helpers.basic import (
+    BaseTest,
+    assertArraysAlmostEqual,
+    assertTensorTuplesAlmostEqual,
+)
 from ..helpers.basic_models import (
     BasicModel_ConvNet,
     BasicModel_MultiLayer,
     BasicModel_MultiLayer_MultiInput,
 )
 from ..helpers.conductance_reference import ConductanceReference
-from ..helpers.utils import (
-    BaseTest,
-    assertArraysAlmostEqual,
-    assertTensorTuplesAlmostEqual,
-)
 
 
 class Test(BaseTest):
