+++ tests/attr/layer/test_layer_deeplift_basic.py
@@ -9,17 +9,17 @@ from torch import Tensor
 
 from captum.attr._core.layer.layer_deep_lift import LayerDeepLift, LayerDeepLiftShap
 
-from ..helpers.basic_models import (
-    BasicModel_MultiLayer,
-    LinearMaxPoolLinearModel,
-    ReLULinearDeepLiftModel,
-)
-from ..helpers.utils import (
+from ...helpers.basic import (
     BaseTest,
     assert_delta,
     assertArraysAlmostEqual,
     assertTensorAlmostEqual,
     assertTensorTuplesAlmostEqual,
+)
+from ..helpers.basic_models import (
+    BasicModel_MultiLayer,
+    LinearMaxPoolLinearModel,
+    ReLULinearDeepLiftModel,
 )
