+++ tests/attr/neuron/test_neuron_ablation.py
@@ -10,12 +10,12 @@ from torch.nn import Module
 from captum._utils.typing import BaselineType, TensorOrTupleOfTensorsGeneric
 from captum.attr._core.neuron.neuron_feature_ablation import NeuronFeatureAblation
 
+from ...helpers.basic import BaseTest, assertTensorAlmostEqual
 from ..helpers.basic_models import (
     BasicModel_ConvNet_One_Conv,
     BasicModel_MultiLayer,
     BasicModel_MultiLayer_MultiInput,
 )
-from ..helpers.utils import BaseTest, assertTensorAlmostEqual
 
 
 class Test(BaseTest):
