+++ tests/attr/neuron/test_neuron_conductance.py
@@ -11,12 +11,12 @@ from captum._utils.typing import BaselineType, TensorOrTupleOfTensorsGeneric
 from captum.attr._core.layer.layer_conductance import LayerConductance
 from captum.attr._core.neuron.neuron_conductance import NeuronConductance
 
+from ...helpers.basic import BaseTest, assertArraysAlmostEqual
 from ..helpers.basic_models import (
     BasicModel_ConvNet,
     BasicModel_MultiLayer,
     BasicModel_MultiLayer_MultiInput,
 )
-from ..helpers.utils import BaseTest, assertArraysAlmostEqual
 
 
 class Test(BaseTest):
