+++ tests/attr/neuron/test_neuron_integrated_gradients.py
@@ -13,15 +13,15 @@ from captum.attr._core.neuron.neuron_integrated_gradients import (
     NeuronIntegratedGradients,
 )
 
+from ...helpers.basic import (
+    BaseTest,
+    assertArraysAlmostEqual,
+    assertTensorTuplesAlmostEqual,
+)
 from ..helpers.basic_models import (
     BasicModel_ConvNet,
     BasicModel_MultiLayer,
     BasicModel_MultiLayer_MultiInput,
-)
-from ..helpers.utils import (
-    BaseTest,
-    assertArraysAlmostEqual,
-    assertTensorTuplesAlmostEqual,
 )
