+++ tests/attr/test_approximation_methods.py
@@ -4,7 +4,7 @@ import unittest
 
 from captum.attr._utils.approximation_methods import Riemann, riemann_builders
 
-from .helpers.utils import assertArraysAlmostEqual
+from ..helpers.basic import assertArraysAlmostEqual
 
 
 class Test(unittest.TestCase):
