+++ tests/attr/test_common.py
@@ -5,7 +5,7 @@ import torch
 from captum.attr._core.noise_tunnel import SUPPORTED_NOISE_TUNNEL_TYPES
 from captum.attr._utils.common import _validate_input, _validate_noise_tunnel_type
 
-from .helpers.utils import BaseTest
+from ..helpers.basic import BaseTest
 
 
 class Test(BaseTest):
