+++ tests/attr/test_deeplift_basic.py
@@ -10,6 +10,12 @@ from torch.nn import Module
 from captum.attr._core.deep_lift import DeepLift, DeepLiftShap
 from captum.attr._core.integrated_gradients import IntegratedGradients
 
+from ..helpers.basic import (
+    BaseTest,
+    assertArraysAlmostEqual,
+    assertAttributionComparision,
+    assertTensorAlmostEqual,
+)
 from .helpers.basic_models import (
     BasicModelWithReusableModules,
     Conv1dDeepLiftModel,
@@ -17,12 +23,6 @@ from .helpers.basic_models import (
     ReLUDeepLiftModel,
     ReLULinearDeepLiftModel,
     TanhDeepLiftModel,
-)
-from .helpers.utils import (
-    BaseTest,
-    assertArraysAlmostEqual,
-    assertAttributionComparision,
-    assertTensorAlmostEqual,
 )
