+++ tests/attr/test_feature_permutation.py
@@ -6,8 +6,8 @@ from torch import Tensor
 
 from captum.attr._core.feature_permutation import FeaturePermutation, _permute_feature
 
+from ..helpers.basic import BaseTest, assertArraysAlmostEqual, assertTensorAlmostEqual
 from .helpers.basic_models import BasicModelWithSparseInputs
-from .helpers.utils import BaseTest, assertArraysAlmostEqual, assertTensorAlmostEqual
 
 
 class Test(BaseTest):
