+++ tests/attr/test_gradient_shap.py
@@ -10,9 +10,9 @@ from captum._utils.typing import Tensor
 from captum.attr._core.gradient_shap import GradientShap
 from captum.attr._core.integrated_gradients import IntegratedGradients
 
+from ..helpers.basic import BaseTest, assertArraysAlmostEqual, assertTensorAlmostEqual
 from .helpers.basic_models import BasicLinearModel, BasicModel2
 from .helpers.classification_models import SoftmaxModel
-from .helpers.utils import BaseTest, assertArraysAlmostEqual, assertTensorAlmostEqual
 
 
 class Test(BaseTest):
