+++ tests/attr/test_guided_grad_cam.py
@@ -9,8 +9,8 @@ from torch.nn import Module
 from captum._utils.typing import TensorOrTupleOfTensorsGeneric
 from captum.attr._core.guided_grad_cam import GuidedGradCam
 
+from ..helpers.basic import BaseTest, assertTensorAlmostEqual
 from .helpers.basic_models import BasicModel_ConvNet_One_Conv
-from .helpers.utils import BaseTest, assertTensorAlmostEqual
 
 
 class Test(BaseTest):
