+++ tests/attr/test_integrated_gradients_classification.py
@@ -9,8 +9,8 @@ from captum._utils.typing import BaselineType, Tensor
 from captum.attr._core.integrated_gradients import IntegratedGradients
 from captum.attr._core.noise_tunnel import NoiseTunnel
 
+from ..helpers.basic import BaseTest, assertTensorAlmostEqual
 from .helpers.classification_models import SigmoidModel, SoftmaxModel
-from .helpers.utils import BaseTest, assertTensorAlmostEqual
 
 
 class Test(BaseTest):
