+++ tests/attr/test_targets.py
@@ -21,8 +21,8 @@ from captum.attr._core.occlusion import Occlusion
 from captum.attr._core.saliency import Saliency
 from captum.attr._core.shapley_value import ShapleyValueSampling
 
+from ..helpers.basic import BaseGPUTest, BaseTest, assertTensorAlmostEqual
 from .helpers.basic_models import BasicModel_MultiLayer
-from .helpers.utils import BaseGPUTest, BaseTest, assertTensorAlmostEqual
 
 
 class Test(BaseTest):
