+++ tests/insights/test_contribution.py
@@ -8,7 +8,7 @@ import torch.nn as nn
 
 from captum.insights import AttributionVisualizer, Batch, FilterConfig
 from captum.insights.features import BaseFeature, FeatureOutput, ImageFeature
-from tests.attr.helpers.utils import BaseTest
+from tests.helpers.basic import BaseTest
 
 
 class RealFeature(BaseFeature):
